Updated contract
