+++ microservices/contract/src/main/kotlin/org/radarbase/appserver/microservices/contract/calls/ProjectServiceContract.kt
@@ -17,7 +17,6 @@
 package org.radarbase.appserver.microservices.contract.calls
 
 import io.ktor.client.HttpClient
-import io.ktor.client.call.body
 import io.ktor.client.engine.cio.CIO
 import io.ktor.client.plugins.HttpTimeout
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
@@ -25,15 +24,15 @@ import io.ktor.client.request.get
 import io.ktor.client.request.post
 import io.ktor.client.request.put
 import io.ktor.client.request.setBody
-import io.ktor.client.statement.HttpResponse
 import io.ktor.http.ContentType
 import io.ktor.http.appendPathSegments
 import io.ktor.http.contentType
 import io.ktor.serialization.kotlinx.json.json
 import kotlinx.serialization.json.Json
 import org.radarbase.appserver.microservices.contract.response.ProxyResponse
-import org.radarbase.appserver.microservices.contract.utils.Utils.normalizedPath
-import org.radarbase.appserver.microservices.contract.utils.Utils.normalizedUri
+import org.radarbase.appserver.microservices.contract.utils.Utils.createProxyFromResponse
+import org.radarbase.appserver.microservices.contract.utils.Utils.getEndpointFromParts
+import org.radarbase.appserver.microservices.contract.utils.Utils.tryProxyRequest
 import org.radarbase.appserver.microservices.core.dto.ProjectDto
 import kotlin.time.Duration.Companion.seconds
 
@@ -133,66 +132,6 @@ object ProjectServiceContract {
                         appendPathSegments("project", projectId)
                     }
                 }.let { response -> createProxyFromResponse(response) }
-            }
-        }
-    }
-
-    suspend fun createProxyFromResponse(resp: HttpResponse): ProxyResponse {
-        val bodyBytes = resp.body<ByteArray>()
-        val contentType = resp.headers["Content-Type"]
-        val location = resp.headers["Location"]
-
-        return ProxyResponse(
-            status = resp.status.value,
-            contentType = contentType,
-            location = location,
-            body = bodyBytes,
-        )
-    }
-
-    fun getEndpointFromParts(baseUrl: String, path: String, prefix: String?): String {
-        val base: String = normalizedUri(baseUrl)
-        val prefixSegment: String = normalizedPath(prefix)
-        val servicePath: String = normalizedPath(path)
-
-        return (base + prefixSegment + servicePath).removeSuffix("/")
-    }
-
-    suspend fun tryProxyRequest(request: suspend () -> ProxyResponse): ProxyResponse {
-        return try {
-            request()
-        } catch (t: Throwable) {
-            when (t) {
-                is io.ktor.client.plugins.HttpRequestTimeoutException,
-                is kotlinx.coroutines.TimeoutCancellationException,
-                is java.net.SocketTimeoutException,
-                    -> {
-                    ProxyResponse(
-                        status = 504,
-                        contentType = "application/json",
-                        body = """{"error":"gateway_timeout","message":"upstream timed out"}""".toByteArray(),
-                    )
-                }
-
-                is java.net.ConnectException -> {
-                    ProxyResponse(
-                        status = 502,
-                        contentType = "application/json",
-                        body = """{"error":"bad_gateway","message":"cannot reach upstream"}""".toByteArray(),
-                    )
-                }
-
-                is io.ktor.client.plugins.ResponseException -> {
-                    try {
-                        val er = t.response
-                        val b = er.body<ByteArray>()
-                        ProxyResponse(status = er.status.value, contentType = er.headers["Content-Type"], body = b)
-                    } catch (_: Throwable) {
-                        ProxyResponse(502, "application/json", body = """{"error":"bad_gateway"}""".toByteArray())
-                    }
-                }
-
-                else -> ProxyResponse(502, "application/json", body = """{"error":"bad_gateway"}""".toByteArray())
             }
         }
     }
